Group utilities into utils.py
